src/requirements.txt
requests
PyYAML
python-keycloak
python-dotenv
boto3